Improve error handling for mem mapped files
We now handle the case where files opened correctly but didn't actually
exist.
This is an issue because we would just be reading garbage anyway at this
point.
But, having said that, maybe we should wait until the app runs and then
we can open those files...

Ok so we should just enter a loop where we wait for the files to open.