Really only open the memmapped files if they exist
Previously I was incorrectly assuming that CreateFile was the only option
to create mapped files.
OpenMappedFile is the way to go.
Obviously in the future we will poll to check when the file becomes
available.
This way we can open the app before the game.
I'll put a loop in tomorrow morning before work.