First attempt at serial data protocol
Each packet is a packed structure of bytes to send
items in a specific order.
The structure will need to be shared between the arduino
and the service projects.
Must deal with disconnections, and restarts.

Currently implemented logic not to send data if session is not live,
but send at least the first byte to say so.