Blink the RPM redline above shiftrpm
This is currently using millis so that if we change the update frequency
in the future, we can still have a pre-determined blinking period.
But that is not necessarily better than just define it as a multiple of
the period, calculated based on the protocol timing, which we need to
know anyway to secure real-time response.