Implement gear indicator patterns
Remove the pattern caching.
Although it might help performance if we are tight on time,
there is still yet another improvement (SPI) which would be much faster
and would not require caching.

Also, changed the initial redline pattern when rpm > shftrpm.
We now switch all leds off first, then on: this way it is much clearer
when the shft rpm has been hit.