Save game session to a known file name format
If no arguments are passed, 'save' will still create a file with a well
known file name pattern YYYYMMDD-HHMMSS.bin
An optional prefix can be passed in to make the name more unique.
This allows the program to be used from a windows shortcut

Additionally, we save version information in the session file, and we
have more helpful usage information.

TODO: must update replay to read the version information and output to
screen before replaying the session in memory.