Refactor ACCSharedMemory.h
Move ACC_CarModelData out of ACCSharedMemory.h into ACCData.[hc]
This more clerly defines what is part of the shared memory and what
is instead data structures to use the shared memory.

SimDisplayProtocol.h is now in Device/include so that it can be found as
a common header between Service and all Devices.
This must be now manually copied for .ino's to compile.
In the future we will use an Atmel compiler which won't require the
manual copy.