Move simdisplay protocol library to a new home
SimDisplayProtocol is now renamed into simdisplay-protocol.
It is hosted at https://github.com/fenegroni/simdisplay-proto
We use arduino-cli to install it directly from github.com
We reference it in both the UnoR3.ino and the MSBuild solution
so that we can rely on one copy of it.
We don't use submodules yet but might do for the ACC data
library because that one might need to be managed
separately.